Add webapi plugin detection
As explained in the readme, webapi-vim[1] is a dependency of
vim-github-comment. To make this obvious for new users, I added some
code to detect if the plugin is available and if it is not, display a
helpful error message.

Since webapi-vim only uses autoload to define its functions, they don't
actually exist until they are called the first time. Therefore, we need
to call one of the functions before checking if it exists. I decided to
go with `webapi#json#decode` since it is already being used by
vim-github-comment.

Fixes #5.

[1]: https://github.com/mattn/webapi-vim